% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerator_in.R
\name{datagenerator.inloop}
\alias{datagenerator.inloop}
\title{This function is used to generate B data batches each time to illustrate the usage of RenewGLM function}
\usage{
datagenerator_in(beta, n, p, B, family, construct, rho, tempdatadir,
  categorical = FALSE, seed = NA)
}
\arguments{
\item{beta}{designed coefficients for simulated data batches, including the intercept}

\item{n}{sample size of the bth data batch}

\item{p}{number of coefficients to estimate including intercept}

\item{B}{the terminal time to get the result}

\item{family}{exponential family of the responses, choices include \code{c("gaussian","binomial","poisson")} }

\item{construct}{structure of covariance matrix for generating the covariate matrix \code{X}}

\item{rho}{the correlation coefficient in the covariance matrix, choices inlcude \code{c("ind", "cs", "ar1")} }

\item{tempdatadir}{the directory for saving a total of B data batches}

\item{categorical}{set to TRUE to let some \code{X} be dichotomized, the default is FALSE}

\item{seed}{set random seed and the default is \code{NA} }
}
\value{
a data matrix containing response vector \code{y} and covariate matrix \code{X}
}
\description{
This function is used to generate B data batches each time to illustrate the usage of \code{RenewGLM.inloop} function
}
