% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenewGLM_in.R
\name{RenewGLM.inloop}
\alias{RenewGLM.inloop}
\title{Renewable GLM function processing data batches internally}
\usage{
RenewGLM_in(B, tempdatadir, type, init = NA, p, intercept = TRUE)
}
\arguments{
\item{B}{index for the terminal data batch}

\item{tempdatadir}{the directory of the streaming datasets, each data batch includes a covariate matrix \code{X} and a response vector \code{y}}

\item{type}{the GLM family you want to fit your data to \code{c("gaussian","binomial","poisson")} }

\item{init}{the initial value for regression coefficients (default is a vector of \code{0})}

\item{p}{number of coefficients to be estimated, including intercept}

\item{intercept}{if intercept is included in the model, default is \code{TRUE}}
}
\value{
coefficient estimates, standard errors and p-values at data batch B
}
\description{
Looping over data batches inside the function and output the final regression coefficients and their standard errors
}
\examples{
N=1000
B=10
p=5
n=N/B
beta<-c(0.2,-0.2,0.2,-0.2,0.2)

tempdatadir<-"~/Desktop/tempdata"
datagenerator_in(beta=beta, n=n, p=p, B=B, family="binomial", construct="cs",
    rho=0.5, tempdatadir=tempdatadir)
RenewGLM_in(B, tempdatadir=tempdatadir, "binomial", p=p, intercept=TRUE)
unlink(tempdatadir)
}
