% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenewGLM_out.R
\name{RenewGLM.outloop}
\alias{RenewGLM.outloop}
\title{Renewable GLM function processing data batches externally}
\usage{
RenewGLM_out(X, y, type, betahat, infomats, intercept, s, phi)
}
\arguments{
\item{X}{covariate matrix for the current data batch}

\item{y}{response vector for the current data batch}

\item{type}{the GLM family you want to fit your data to \code{c("gaussian","binomial","poisson")} }

\item{betahat}{the old estimates that need to be updated}

\item{infomats}{the old cumulative information matrix that need to be updated}

\item{intercept}{if an intercept is included in the model}

\item{s}{the cumulative sample size (only needs to be specified in Gaussian model, does not include the samples in the current data batch)}

\item{phi}{the old estimate of the dispersion parameter in Gaussian model}
}
\value{
updated coefficient estimates and the cumulative information matrix
}
\description{
Take in data batches sequentially and update the regression coefficients and their standard errors
}
\examples{
N=1000
B=10
p=5
n=N/B
beta<-c(0.2,-0.2,0.2,-0.2,0.2)
infomats<-diag(0,p,p);
betahat<-rep(0,p)
for(b in 1:10){
 data<-datagenerator_out(beta,b,n,"binomial","cs",0.5)
 y<-data[,1]
 X<-data[,-1]
 summary<-RenewGLM_out(X,y,"binomial",betahat,infomats,intercept=TRUE,s,phi)
 betahat<-summary[[1]]
 infomats<-summary[[2]]
 rm(data)
}
sd<-sqrt(diag(solve(infomats)))
pvalue<-2*pnorm(-abs(betahat)/sd)
result<-cbind(betahat=betahat,sd=sd,pvalue=pvalue)
colnames(result)<-c("Estimates","Std.Errors","p-values")
}
